Added helper functions to modify domain
