+++ src/Makevars
@@ -3,8 +3,8 @@ PKG_CPPFLAGS=-I. -I/usr/local/lib/libminizinc/include
 PKG_LDFLAGS=-lmzn
 PKG_LIBS=-L/usr/local/lib/libminizinc	-lmzn 
 OBJECTS.tests=cpp_tests/test-runner.o cpp_tests/test-mzn_parse.o cpp_tests/test-set_params.o cpp_tests/test-mzn_eval.o cpp_tests/test-sol_parse.o  cpp_tests/test-getMissingPars.o cpp_tests/test-deleteItem.o
-OBJECTS.sources=RcppExports.o set_params.o mzn_parse.o mzn_eval.o sol_parse.o getMissingPars.o  deleteItem.o modifyDomain.o
-OBJECTS.helpers=filetoString.o helper_parse.o expVarNames.o helper_sol_parse.o pathStringcheck.o
+OBJECTS.sources=RcppExports.o set_params.o mzn_parse.o mzn_eval.o sol_parse.o getMissingPars.o  deleteItem.o modifyDomain.o modifyConstraint.o
+OBJECTS.helpers=filetoString.o helper_parse.o expVarNames.o helper_sol_parse.o pathStringcheck.o helper_modify.o
 OBJECTS = $(OBJECTS.sources) $(OBJECTS.tests) $(OBJECTS.helpers) $(SOURCES:.cpp=.o)
 strippedLib: $(SHLIB)
 	if test -e "/usr/bin/strip"; then /usr/bin/strip --strip-debug $(SHLIB); fi
